feat: Started adding DispatchAreas

+++ src/Application/Handler/DeleteStateHandler.php
@@ -3,6 +3,7 @@
 namespace App\Application\Handler;
 
 use App\Application\Contract\HandlerInterface;
+use App\Application\Exception\StateNotEmptyException;
 use App\Domain\Command\DeleteStateCommand;
 use App\Domain\Event\StateDeleted;
 use App\Domain\Repository\StateRepositoryInterface;
@@ -24,6 +25,10 @@ class DeleteStateHandler implements HandlerInterface
     public function __invoke(DeleteStateCommand $command): void
     {
         $state = $this->stateRepository->getById($command->getId());
+
+        if (!$state->getDispatchAreas()->isEmpty()) {
+            throw new StateNotEmptyException('Cannot delete Area as it is not empty');
+        }
 
         $this->stateRepository->delete($state);
